Developement

+++ research/views.py
@@ -36,6 +36,9 @@ class ResearchApiViewList(APIView):
     
     def post(self, request):
         """Create a new research object"""
+        if not request.user.is_authenticated:
+            return  Response({"error": "User is not authenticated"}, status = status.HTTP_400_BAD_REQUEST)
+
         serializer = self.serializer_class(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -47,18 +50,37 @@ class ResearchApiViewList(APIView):
         logger.error(f'\n\n{serializer.errors}\n')
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+
 class ResearchApiViewDetail(APIView):
 
     serializer_class = ResearchSerializer
     permissions_classes = [IsAuthenticated]
 
+    def get_object(self, pk):
+        try:
+            return Research.objects.get(pk=pk)
+        except Research.DoesNotExist:
+            raise Http404
+
     def get(self, request, researchId):
-        print(request.user)
-        """Teturn Research detail for requested research"""
-        research = Research.objects.get(id =researchId)
+        """Return Research detail for requested research"""
+
+        if not request.user.is_authenticated:
+            return  Response({"error": "User is not authenticated"}, status = status.HTTP_400_BAD_REQUEST)
+        research = self.get_object(pk =researchId)
         serializer = self.serializer_class(research)
         response_status = status.HTTP_200_OK
         return  Response({"data": serializer.data}, status = response_status)
+
+    
+    def delete(self, request, pk, format=None):
+
+        if not request.user.is_authenticated:
+            return  Response({"error": "User is not authenticated"}, status = status.HTTP_400_BAD_REQUEST)
+        
+        research = self.get_object(pk)
+        research.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
 
 class ParticipantList(APIView):
     """A view class to manage participants"""
@@ -78,6 +100,9 @@ class ParticipantList(APIView):
     
     def post(self, request):
         """Create a new Participant"""
+        if not request.user.is_authenticated:
+            return  Response({"error": "User is not authenticated"}, status = status.HTTP_400_BAD_REQUEST)
+
         serializer = self.serializer_class(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -98,13 +123,18 @@ class ParticipantDetails(APIView):
             raise Http404
 
     def get(self, request, pk, format=None):
+        if not request.user.is_authenticated:
+            return  Response({"error": "User is not authenticated"}, status = status.HTTP_400_BAD_REQUEST)
+
         participant = self.get_object(pk)
         serializer = self.serializer_class(participant)
         return Response(serializer.data)
     
     def patch(self, request, pk, format=None):
+        if not request.user.is_authenticated:
+            return  Response({"error": "User is not authenticated"}, status = status.HTTP_400_BAD_REQUEST)
+
         participant = self.get_object(pk)
-        print(participant, request.data)
         serializer = self.serializer_class(participant, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
@@ -112,6 +142,9 @@ class ParticipantDetails(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def put(self, request, pk, format=None):
+        if not request.user.is_authenticated:
+            return  Response({"error": "User is not authenticated"}, status = status.HTTP_400_BAD_REQUEST)
+
         Participant = self.get_object(pk)
         serializer = self.serializer_class(Participant, data=request.data)
         if serializer.is_valid():
@@ -120,6 +153,10 @@ class ParticipantDetails(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk, format=None):
+        
+        if not request.user.is_authenticated:
+            return  Response({"error": "User is not authenticated"}, status = status.HTTP_400_BAD_REQUEST)
+
         Participant = self.get_object(pk)
         Participant.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
@@ -138,6 +175,7 @@ class NetworkAPIView(APIView):
             logger.error("User is not authenticated")
             return  Response({"error": "User is not authenticated"}, status = status.HTTP_400_BAD_REQUEST)
 
+        round = request.query_params.get('round', None)
         error = None
         centrality = None
         density = None
@@ -150,6 +188,14 @@ class NetworkAPIView(APIView):
         serializer = self.serializer_class(research)
         interactions =serializer.data.get('interactions') 
 
+
+        if round:
+            tmp = []
+            for interaction in interactions:
+                if interaction['round'] == int(round):
+                    tmp.append(interaction)
+            interactions = tmp
+
         """Get list of research interactions"""
         response_status = status.HTTP_200_OK
         logger.info("Creating a network from all interaction for current research")
@@ -157,35 +203,35 @@ class NetworkAPIView(APIView):
         edges = list(network.edges)
         nodes = list(network.nodes)
 
-        if request.query_params.get('centrality', None):
+        if request.query_params.get('centrality', None) and network:
             try:
                 centrality = Network.calculate_betweens(network)
             except Network.nx.NetworkXException as e:
                 logger.error(f'Could not calculate betweens for network\n{e}')
                 error = e.message
 
-        if request.query_params.get('density', None):
+        if request.query_params.get('density', None) and network:
             try:
                 density = Network.calculate_density(network)
             except Network.nx.NetworkXException as e:
                 logger.error(f'Could not calculate density for network\n{e}')
                 error = e.message
 
-        if request.query_params.get('radius', None):
+        if request.query_params.get('radius', None) and network:
             try:
                 radius = Network.calculate_radius(network)
             except Network.nx.NetworkXException as e:
                 logger.error(f'Could not calculate radius for network\n{e}')
                 error = e.message
 
-        if request.query_params.get('diameter', None):
+        if request.query_params.get('diameter', None) and network:
             try:
                 diameter = Network.calculate_diameter(network)
             except Network.nx.NetworkXException as e:
                 logger.error(f'Could not calculate diameter for network\n{e}')
                 error = e.message
 
-        if request.query_params.get('reciprocity', None):
+        if request.query_params.get('reciprocity', None) and network:
             try:
                 reciprocity = Network.calculate_reciprocity(network)
             except Network.nx.NetworkXException as e:
